Conflicts:
	Mortal Kombat/src/model/Capa.cpp
	Mortal Kombat/src/model/CapaFondo.cpp
	Mortal Kombat/src/model/Mundo.cpp
	Mortal Kombat/src/view/LTexture.cpp
	Mortal Kombat/src/view/Sprite.cpp

+++ Mortal Kombat/src/view/Frame.cpp
@@ -8,7 +8,7 @@
 
 #include "Frame.h"
 
-Frame::Frame(int x,int y,int alto,int ancho){
+Frame::Frame(float x,float y,float alto,float ancho){
 	X = x;
 	Y = y;
 	Alto = alto;
@@ -16,5 +16,4 @@ Frame::Frame(int x,int y,int alto,int ancho){
 }
 
 Frame::~Frame(){
-	//Liberar
 }
